# docs/swagger.yaml
basePath: /
definitions:
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
  handler.getAllAccessGroupByQueryResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/tc.AccessGroup'
        type: array
    type: object
...
  handler.statusResponse:
    properties:
      status:
        type: string
    type: object
  tc.AccessGroup:
    properties:
      access_group:
        type: string
      access_status:
        type: string
      client_status:
        type: integer
      clsrv:
        type: string
      created_at:
        type: string
      dognum:
        type: string
      id:
        type: string
      iface_desc:
        type: string
      iface_host:
        type: string
      iface_name:
        type: string
      in_policy:
        type: string
      ip:
        type: string
      out_policy:
        type: string
      router_name:
        type: string
      updated_at:
        type: string
    type: object
...
  tc.FilterAccessGroup:
    properties:
      _:
        type: string
      __:
        type: string
      filter_desc:
        type: string
      filter_type:
        type: string
      filter_val:
        type: string
      router_name:
        type: string
      user_id:
        type: string
    required:
    - filter_type
    - filter_val
    - router_name
    type: object
...
  tc.UpdateAccessGroupInput:
    properties:
      access_status:
        type: string
      access_status_old:
        type: string
      id:
        type: string
      msg:
        type: string
    required:
    - access_status
    - id
    type: object
...
host: tc.kaztranscom.kz:9000
info:
  contact:
    email: 7773785631@mail.ru,K.Yatsenko@kaztranscom.kz
    name: Серышев А.В., Яценко К.Ю.
  description: API Сервер для Программы Total Control
  title: Total Control
  version: "1.0"
paths:
  /api/access_group:
    get:
      consumes:
      - application/json
      description: Список данных Access Group по переданным параметрам
      operationId: get-access-group-by-query
      parameters:
      - description: в формате yyyy-mm-dd
        in: query
        name: date
        type: string
      - description: Учитывать или нет фильтр
        in: query
        name: check_filter
        type: string
      - description: Закрыто, В обработке
        in: query
        name: access_status
        type: string
      - description: 'можно указать несколько, через запятую '
        in: query
        name: router_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllAccessGroupByQueryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Список данных
      tags:
      - Access Group
    post:
      consumes:
      - application/json
      description: Добавление в базу новых данных
      operationId: insert-access-group
      parameters:
      - description: Данные для внесения в базу
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/tc.AccessGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Добавление данных
      tags:
      - Access Group
  /api/access_group/{id}:
    put:
      consumes:
      - application/json
      description: Редактирование информации
      operationId: update-access-group
      parameters:
      - description: uuid для редактирования
        in: path
        name: id
        required: true
        type: string
      - description: Данные для редактирования
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/tc.UpdateAccessGroupInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Редактирование
      tags:
      - Access Group
  /api/access_group/access_group_history:
    post:
      consumes:
      - application/json
      description: Получение  данных, история изменений, разные данные
      operationId: get-access-group-history
      parameters:
      - in: query
        name: id
        type: string
      - in: query
        items:
          type: string
        name: ids
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получение данных
      tags:
      - Access Group
  /api/access_group/download:
    get:
      consumes:
      - application/json
      description: Скачивание файла статистики Форма 1(Excell)
      operationId: download-excell-file-access-group
      parameters:
      - description: в формате yyyy-mm-dd
        in: query
        name: date
        type: string
      - description: Учитывать или нет фильтр
        in: query
        name: check_filter
        type: string
      - description: Закрыто, В обработке
        in: query
        name: access_status
        type: string
      - description: 'можно указать несколько, через запятую '
        in: query
        name: router_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Скачивание файла списка Access Group
      tags:
      - Access Group
  /api/access_group/filter_access_group:
    get:
      consumes:
      - application/json
      description: Получение  данных, Фильтры, разные данные
      operationId: get-filter-access-group
      parameters:
      - in: query
        name: created_at
        type: string
      - in: query
        name: filter_desc
        type: string
      - in: query
        name: filter_type
        type: string
      - in: query
        name: filter_val
        type: string
      - in: query
        name: id
        type: string
      - in: query
        name: router_name
        type: string
      - in: query
        name: user_id
        type: string
      - in: query
        name: user_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получение данных
      tags:
      - Filter Access Group
    post:
      consumes:
      - application/json
      description: Добавление в базу новых данных, история изменений, разные данные
      operationId: insert-filter-access-group
      parameters:
      - description: Данные для внесения в базу
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/tc.FilterAccessGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Добавление данных в filter(Исключения)
      tags:
      - Filter Access Group
  /api/access_group/filter_access_group/{id}:
    delete:
      consumes:
      - application/json
      description: Удаление Фильтра
      operationId: delete-filter-access-group
      parameters:
      - description: uuid фильтра
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Удаление
      tags:
      - Filter Access Group
...
  /api/router_onyma/filter_router_onyma/{id}:
    delete:
      consumes:
      - application/json
      description: Удаление Фильтра
      operationId: delete-filter
      parameters:
      - description: uuid фильтра
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Удаление
      tags:
      - Filter Router Onyma
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"